Add a job that creates a default Admin Set for a tenant
After a long discussion with @atz, @cbeer, and @jcoyne, I have decided
to start by adding a new job to create a default Admin Set. A follow-up
PR might be to sacrifice idempotency to handle account creation in a
single job. This would be temporary, to ensure that default AS creation
works during the community beta testing period (May 15th through
June). Ultimately we should pursue the "UX step" approach outlined by
@atz in https://github.com/projecthydra-labs/hyku/pull/884, though we
likely don't have time to complete and test that for a month or two.