restore pals-specific CollectionPresenter overrides
We cherry-picked this commit from Hyku: ac8f314e67a3b6edc9965877c9f2911c1c88b4a7

In that commit, the CollectionPresenter #class_eval override was replaced with a decorator. The decorator was Hyku's though, and thus didn't have the additional pals-specific overrides. We restore those in this commit