now that analytics are set per tenant, we do not need the collections controller override
this is because each tenant has to set up analytics for themselves. the override did a search in the tenant db to see if the collection could be found and if not, it did not show.
that was valuable when all tenants shared the same analytics profile, to make sure we got info for that tenant specifically. now, it is unnecessary