408-for-the-last-time (#494)
* make sure that the account analytics settings are being set.

also, add more hint details

* disable Style/RedundantSelf for #reload_analytics

* guard against the account setting value being an empty string.

if an account once had analytics settings that have been removed, the value for each would be an empty string. when using the or/equal operator with an empty string  and non empty string, the empty string is returned. we cannot however check for the presence of a variable while also using the or/equal operator.