🎁 Allowing per tenant ability to disable IIIF Print
The current (as of <2023-08-10 Thu>) implementation of IIIF Print
assumes that we generate derivatives at the application level.  However,
we want options for generating at the tenant level.

As noted, I have chosen to default using IIIF print to `true` because
that was the behavior before this commit.

In other words, this commit is functionally a non-change to
the production code-base (assuming no one goes and flips some switches).

Related to:

- https://github.com/scientist-softserv/palni-palci/issues/656
- https://github.com/scientist-softserv/palni-palci/issues/657
- https://github.com/scientist-softserv/palni-palci/issues/658
- https://github.com/scientist-softserv/palni-palci/issues/659