♻️ Refactor and Test Sushi Date Boundary Conditions
This commit contains two primary changes:

1. Favoring objects responding to `#as_json`
2. Date range boundaries for Sushi report dates

**`#as_json`**

In Rails, when an object responds to `#as_json`, that object can be
directly passed to the `render json: object` directive, and Rails will
"figure it out".

This refactor also adds a spec to the Sushi::PlatformUsageReport.
Having a single spec will help us with further code refactoring as we
discover common behavior.

**Date Range Boundaries**

In discussions with team members, the date range boundaries happen in
month chunks.  I have taken time to write specs whose
documentation (e.g. the strings after the `it` and `context` and
`describe`) read as the intended behavior.

I have also added comments to explain the intention.  Please read these
carefully and flag any inconsistencies.  Date range logic is tricky,
especially considering the constraints of both Sushi's expected
interface and the reality of having delays in reporting logic.

Ultimately, the goal of these date ranges is for us to inform the
consumer of the API about the complete months of data that we have.

**Miscellaney**

It also adresses a type, namely that path to one of the API end-points
did not have a leading "/"