🐛 Ensure consistent logic path for when iiif viewer
There is a conflict between the `IiifPrint` and `Hyrax::IiifAv` gems;
namely they both have strong opinions about how to sniff out if we
should use the `iiif_viewer?`.  Compounding this, is that IiifPrint
decorates `Hyrax::WorkShowPresenter` and Hyku extends
`Hyrax::WorkShowPresenter` then includes
`Hyrax::IiifAv::DisplaysIiifAv`.

The end result is that the logic to determine if we should show pages
split from the PDF is never called.  Yet, if we were to solely use
IiifPrint we'd ignore rendering audio and vidoe in the iiif viewer.

So this commit peels that back so that we're using the logic (brought
forward by IiifPrint) but ensuring our Hyku presenters are using that
logic.

Why move the logic out of the Hyku instance and into a module?  Because
that module contains the per-tenant antics of IiifPrint and its PDF
relationship.

There is larger work to do in regards to incorporating this logic into
IiifPrint and the Hyrax::IiifAv gem.

Does this work?  Please pull down a branch and check.  I'm at the end of
my day (and then some).

Related to:

- https://github.com/scientist-softserv/palni-palci/issues/659