Use value labels in Contributing Library facet
Contributing Library is a controlled field. Before this change, its facet was displaying the ID of the term. Now, the facet will display each controlled value's human-readable label.

contributing_library_sim is the Solr field that is used for the facet, whereas contributing_library_tesim is the "main" Solr field that maps to the form input, displays on the SolrDocument, etc. This means we can safely override how contributing_library_sim is indexed without being concerned about unintentionally changing how the values display elsewhere in the app.