refactor the item report errors
instead of always throwing a 422 error for having no results, we will throw a 422 only if the parameter given is invalid. e.g., the item_id does not exist in the Hyrax::CounterMetric table.

if there are no results because of the date range given however, we will throw a 404 and inform the user of that.