create validations for the `access_method` query param.
if the `access_method` query param is passed in with at least one valid option, we return the valid option(s) to be reflected in the `report filters` array.
if no valid options are passed through the param, we throw a `Sushi::InvalidParameterValue` error.

either way, the `access_method` that shows in the `item_report` is hard coded to say `Regular` because thats the only option we are allowing at this time. if that changes, the code in Sushi::ItemReport#report_items will need to be updated.