📚 Add description to models that are abbreviations
The OER, CDL, and ETD models are all abbreviations.  This commit adds a
human readable name to this as well as a link to the wikidata.org node
that describes the resource.

This work is in part my read through of the models in an effort to find
the appropriate properties for `year of publication` and `author`; both
fields required for the ongoing Sushi work.

Tangentially related to:

- https://github.com/scientist-softserv/palni-palci/issues/721
- https://github.com/scientist-softserv/palni-palci/issues/720