🚧 Pass search term to PDF.js
This commit will add `params[:q]` to the show page URL so we can pass it
to the PDF.js viewer to leverage the auto search functionality.  Also,
abstracted the viewer URL to a helper method using Lafayette College
Libraries' Hyrax application as a reference.

To enable catalog searching we also have to index the text found in the
PDFs onto the file set.  We will leverage the poppler-utils package
(`pdftotext`) to extract the text from the PDFs.

Ref:
  - https://github.com/LafayetteCollegeLibraries/spot/pull/14