🎁 Add PDF.js viewer to Flipflop
This commit will allow admins to toggle between using the new default
PDF.js viewer or the Universal Viewer with IIIF Print for PDF rendering.
Using a custom label of `PDF.js|UV` for the feature makes it more clear
than simply using the default `on|off` labels.  To enable this visual we
need to override the Flipflop's `StrategiesController` and reference our
helper's constant.  This should allow for easy extensions in the future.
This commit also modify's all the previous logic of `:use_iiif_print` to
the new `:default_pdf_viewer` feature.