🐛 Bring over Hyrax::IiifHelper
There seems to be a bug in IIIF Print where its version of the
IiifHelper is not getting loaded.  This is a quick fix to just bring in
the Hyrax version and adding the `#uv_search_param` method to avoid a
`NoMethodError` on the work show page when using the Universal Viewer.