Match logic for download buttons
This commit will match the PDF download button logic with the actions
download button logic.  Also this will move the Flipflop check outside
of the download_pdf partial (in a helper) to the render call on the
show.html.erb so we don't have an empty div if it fails that check.  We
now pass in the file_set_id instead of doing the logic in the partial as
well to keep the partial cleaner.