🚧 Remove previous `Download PDF` button
This commit will remove the `Download PDF` button away from the
thumbnail rendering so we can set it up for better controls later
because the admin wants to be able to toggle the button on and off just
like the viewer.  Also, the logic for displaying the viewer has been
cleaned up.  We will need to copy this logic over to the other themes.