🧹 Fix a bug and some styling
Fixed a bug where the old helper method was still being used.  Removed
the download button from reshare theme.  Gave the download button a
little more room on the top and bottom.  I couldn't figure out how to
target the styling from the scss so I ended up doing it in the
button_tag helper inline.  I left it there just in case it's my
environment or I'm just doing something silly.  We can address that in a
later PR if we want.