♻️ Modify JSON response message to conform to spec
The [spec defines numerous exceptions][1].  This refactor moves towards
conforming to those specifications.

This change also builds on raising exceptions when data is not yet
ready.  To then guard against those exceptions, we need to either rescue
in the JSON document (which would be rather gross) or wrap the Sushi
methods with a guarding function.  I chose the latter.

See the inline comments for further discussion/documentation on these
changes.

Related to:

- https://github.com/scientist-softserv/palni-palci/issues/688

[1]: https://countermetrics.stoplight.io/docs/counter-sushi-api/sgmqulr0emsi6-exception