:gift: Default sort works by title for Collection Show Page
In this commit we introduce and override because default hyrax title indexes as title_ssi. This will not work for PALS since they index as title_ssi.

Note: This will fix the default behavior however the sorting is still not working. This may be taken care of in a separate issue.

Issue: https://github.com/scientist-softserv/palni-palci/issues/800