🐛 Bring back missing download link on pdf
This commit will fix a bug where the Download PDF link was erroneously
removed in a previous commit.  It was originally removed because it was
rendering on the work show page but in doing so it was also removed from
the file set show page.  We now bring it back with a conditional that
only renders it on the file set show page.

Ref
  - https://github.com/scientist-softserv/palni-palci/issues/814
  - https://github.com/scientist-softserv/palni-palci/pull/788