🐛 Raise 3060 Sushi Exception if | provide for author field
According to the Sushi spec, the [Item Report][1] does not account for multiple authors.

> Identifier of a specific author usage is being requested for. If
> omitted, all items on the platform with usage for the customer will be
> returned.

Later, for `yop` there is language about how to encode multiple values.

Which, I believe means that if we include a `|` in the author we should
return a 3060.  So I'll set about working on that.  (That does assume a
person does not have a `|` in their name).  Note, the data encoding
we're doing is to have pipe separation of multiple authors in the
database.  The record you're seeing has `|Summer|Winter|` for a
multi-author file.  And there's a direct match.

Related to:

- https://github.com/scientist-softserv/palni-palci/issues/721

[1]: https://countermetrics.stoplight.io/docs/counter-sushi-api/5a6e9f5ddae3e-ir-item-report