update filter handling for the platform usage report
`Searches_Platform` is a valid metric_type filter according to
https://countermetrics.stoplight.io/docs/counter-sushi-api/xggfvq5by3gx0-pr-p1-report-filters.

before this commit we were always returning `Searches_Platform`, despite
whether a `metric_type` filter was being used.

with this commit, we only return `Searches_Platform` if the `metric_type`
param includes it, or is not being passed at all.

ref: https://github.com/scientist-softserv/palni-palci/issues/687#issuecomment-1736117022