🎁 PDF file sets now selectable in file manager
Currently, the file manager isn't selecting between PDFs for display in
the viewer.  If there are multiple PDFs, changing the representative
media/thumbnail doesn't affect which one displays in the viewer.  This
commit will add the ability for the user to change which PDF is being
displayed through the file manager.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/839