🩹 Temporary fix for video processing
This commit will set the IIIF manifest URI for videos to be the Hyrax
download video.  This is a temporary fix until we can get the video
processing speeds faster.  It won't make a difference if the video that
was ingested was already processed, however, this workaround falls apart
if the uploaded video is a raw video file where the size could be
enormous.  That is not ideal for streaming.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/853