🎁 Add ability to sort dashboard collections
This commit will add the ability to sort collections on the dashboard.
The collections work differently than the works so we have to make sure
the sort field is added to the search builder.  Because when you click
into a collection can view the works attached to the collection, we have
to approach it differently because otherwise the sort fields from the
dashboard index page will also be applied to the show page.  We decorate
Blacklight helpers to check the controller action name as our
conditional.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/748