🐛 Fix user despositing to specific admin sets
This commit will add back the ability for everyone to be able to create
a work.  This is one of the conditionals that need to be met for a user
to deposit a work to an admin set.  The other is already being met,
which is that the user has to have either a depositor role for that
specific admin set or in a group that has a depositor role for that
specific admin set.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/864