🐛 Omit DataType=Platform when provided DataType does not include Platform
Prior to this commit, if the query parameters included only non-platform
data_type (e.g. article) then we would continue to show the platform
data type node.

With this commit, we skip showing that element if we have been provided
explicit data_type(s) and none of them are Platform.

Related to:

- https://github.com/scientist-softserv/palni-palci/issues/686#issuecomment-1785326034