🐛 add limited reindex to collections job
Add limited reindex to the collection_id  path and refactor in honor of 'Don't repeat yourself'.
Before this commit calling this job from a rake task took hours to complete.
It may be because of nested indexing. This change will ensure limited reindexing
is enforced regardless of the logical branch