🐛 Favor overriding CollectionsController
This commit will allow us to override the CollectionsController instead
of the Blacklight::ConfigurationHelperBehavior.  A bug was noticed when
switching the sort fields that the field selected wouldn't stick in the
UI even though the fields were actually sorting correctly.  Overriding
the CollectionsController will be a better pattern since we override the
WorksController in the same way.  The reason why we didn't do that in
the first place was because it caused the show page to have the same
sort fields as the index page.  In this commit, we are also modifying
the show page to reconfigure the sort fields to address the issue.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/748