:gift: Add HYRAX_USR_SOLR_GRAPH_NESTING env variable
Since PALS is on 3.6 version of hyrax it has graph_indexer to improve the speed of queries and reindexing. However PALS does not have the env var set to enable it in the various environments. There was an ENV['HYRAX_USE_SOLR_GRAPH_SETTING'] being set in hyrax.rb, however it doesn't look like it was being honored. See screenshot of error message from production.