🐛 Fix top level collections
Fixing the app to use the graph indexer means that the query for
top level collections will not work. It relied on an index field,
`nesting_collection__deepest_nested_depth_isi` which was added by the
nesting indexer. With the use of the graph indexer, this term no longer
exists.

A simpler solution is to identify a top level collection as one which
is not a member of any other collection.

Along with this change, come code duplication is also being cleaned up.