🎁 Add toggle for pdf viewer and download button
This commit will add two properties for all work types to allow for the
pdf viewer and download button to be toggled on and off on the work show
page.  These two options will exist on the forms for each work type on
the files tab.  This will also hide the default download button on
PDF.js so the users can only download from the newly introduced download
button.