🐛 Fix mime type for IIIF manifest
This commit will use `audio/mpeg` as the mime type for mp3 resources in
the IIIF manifest instead of relying on the characterized mime type of
the original file.  This is a problem when our file is a wav file and it
gets characterized as `audio/x-wave` which doesn't work with Universal
Viewer.  Instead, since we are serving up the mp3 derivative in the
manifest, we should use `audio/mpeg` as the mime type.  We also remove
the ogg format because it is not being used and having that hard coded
as `audio/mpeg` as well feels incorrect.  This however doesn't solve the
bigger problem of the File Set not gettig the original file in the first
place but that is a separate issue and would need some investigation.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/887