🎁 Add search form to splash page
This feature introduces a search on the proprietor's home page (e.g. the
root host of the Hyku instance).

That form will then submit a search against the configured host for
cross-tenant searching.

In the case of "commonsarchive.org", that will be
"//search.commonsarchive.org/catalog" (as demonstrated in the hard-coded value).

Related to:

- https://github.com/scientist-softserv/palni-palci/issues/947