create child job class to use new Sidekiq queue
There is no way to reassign a job to another queue once it has been
created. To use the new :resource_intensive Sidekiq queue, we need a
separate job class that is configured to use that queue.

Enter CreateLargeDerivativesJob.

It functions exactly the same as CreateDerivativesJob, except that it
queues into the :resource_intensive queue and has a higher priority.
This way, we can have two separate job classes with different Sidekiq
configurations while using the same application logic for both.