🎁 Grant group access for users to a resource
The previous approach was to grant a user access to each work and their
child works and their file sets, which can be a lot of trips to the
database.  This commit will instead grant the user read access to the
group that the work, its child works, and all of its file sets belong
to which will reduce the number of trips to the database.