Rake task for updating collection_type_gid
The structure of the stored collection type gid changed in Hyrax 3.0

A rake task is needed to update the collection_type_gid that is stored
on the collection as part of the upgrade from Hyrax 2.x process. This
references the logic from the rake task that Hyrax has in versons 3 & 4,
and updates it to work within Hyku tenants.

The definition of collection_type_gid= changed in Hyrax 5.0. This
rake task is known to work for updates to versions prior to Hyrax 5.
Functionality for later versions is unknown.