🐛 Fix date indexing to return more than YYYY
Prior to this commit when the date_created property was indexed, only
the year was indexed.  If the input was something like "2024-01-01" then
`date_ssi` would only be "2024".  This commit will index YYYY-MM-DD,
YYYY-MM, and YYYY if given.  Anything other than those three formats
will be indexed as is and won't work correctly with sorting or range
queries.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/973