🧹 Hide custom Hyku Commons workflow
This commit will hide the custom Hyku Commons workflow from the admin
workflow form.  This is a temporary measure to ensure that the option is
not going to show up in the admin interface to reduce confusion.  The
long-term fix will be to create a script and delete the custom workflow
from the database from each tenant, however it was warned that working
with Siptiy::Workflow might be a little tricky.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/448